test: BlobText取り込みcontractを分離する
